Rename partials to query partials
