+++ lib/next-api/wordpress/comments/mutationAddComment.js
@@ -1,5 +1,5 @@
 import {nextApiRoutes} from '@/lib/next-api/connector'
-import commentsFields from '@/lib/wordpress/_partials/commentsFields'
+import commentsFields from '@/lib/wordpress/_query-partials/commentsFields'
 import {gql} from '@apollo/client'
 
 // Mutation: Add a comment to the given post.
