+++ lib/wordpress/_query-partials/archivePageInfo.js
