+++ lib/wordpress/_query-partials/globalPostFields.js
