+++ lib/wordpress/comments/mutationInsertComment.js
@@ -1,4 +1,4 @@
-import commentsFields from '@/lib/wordpress/_partials/commentsFields'
+import commentsFields from '@/lib/wordpress/_query-partials/commentsFields'
 import {gql} from '@apollo/client'
 
 // Mutation: Add a comment to the given post.
