+++ lib/wordpress/comments/queryCommentsByPostId.js
@@ -1,5 +1,5 @@
-import archivePageInfo from '@/lib/wordpress/_partials/archivePageInfo'
-import commentsFields from '@/lib/wordpress/_partials/commentsFields'
+import archivePageInfo from '@/lib/wordpress/_query-partials/archivePageInfo'
+import commentsFields from '@/lib/wordpress/_query-partials/commentsFields'
 import {gql} from '@apollo/client'
 
 // Query: retrieve comments by post databaseId.
