+++ lib/wordpress/pages/queryDefaultPageData.js
@@ -1,4 +1,4 @@
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
 import {gql} from '@apollo/client'
 
 // Query: retrieve default SEO and other page data.
