+++ lib/wordpress/pages/queryError404Page.js
@@ -1,4 +1,4 @@
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
 import {gql} from '@apollo/client'
 import {singlePageFragment} from './queryPageById'
