+++ lib/wordpress/posts/queryPostById.js
@@ -1,11 +1,11 @@
-import authorPostFields from '@/lib/wordpress/_partials/authorPostFields'
-import categoriesPostFields from '@/lib/wordpress/_partials/categoriesPostFields'
-import commentsPostFields from '@/lib/wordpress/_partials/commentsPostFields'
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
-import featuredImagePostFields from '@/lib/wordpress/_partials/featuredImagePostFields'
-import globalPostFields from '@/lib/wordpress/_partials/globalPostFields'
-import seoPostFields from '@/lib/wordpress/_partials/seoPostFields'
-import tagsPostFields from '@/lib/wordpress/_partials/tagsPostFields'
+import authorPostFields from '@/lib/wordpress/_query-partials/authorPostFields'
+import categoriesPostFields from '@/lib/wordpress/_query-partials/categoriesPostFields'
+import commentsPostFields from '@/lib/wordpress/_query-partials/commentsPostFields'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
+import featuredImagePostFields from '@/lib/wordpress/_query-partials/featuredImagePostFields'
+import globalPostFields from '@/lib/wordpress/_query-partials/globalPostFields'
+import seoPostFields from '@/lib/wordpress/_query-partials/seoPostFields'
+import tagsPostFields from '@/lib/wordpress/_query-partials/tagsPostFields'
 import {gql} from '@apollo/client'
 
 // Fragment: retrieve single post fields.
