+++ lib/wordpress/tags/queryPostsByTag.js
@@ -2,8 +2,8 @@ import {
   archivePostFragment,
   archivePosts
 } from '@/lib/wordpress/posts/queryPostsArchive'
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
-import seoPostFields from '@/lib/wordpress/_partials/seoPostFields'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
+import seoPostFields from '@/lib/wordpress/_query-partials/seoPostFields'
 import {gql} from '@apollo/client'
 
 // Query: retrieve posts tag archive.
