+++ lib/wordpress/teams/queryTeamById.js
@@ -1,8 +1,8 @@
-import authorPostFields from '@/lib/wordpress/_partials/authorPostFields'
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
-import featuredImagePostFields from '@/lib/wordpress/_partials/featuredImagePostFields'
-import globalPostFields from '@/lib/wordpress/_partials/globalPostFields'
-import seoPostFields from '@/lib/wordpress/_partials/seoPostFields'
+import authorPostFields from '@/lib/wordpress/_query-partials/authorPostFields'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
+import featuredImagePostFields from '@/lib/wordpress/_query-partials/featuredImagePostFields'
+import globalPostFields from '@/lib/wordpress/_query-partials/globalPostFields'
+import seoPostFields from '@/lib/wordpress/_query-partials/seoPostFields'
 import {gql} from '@apollo/client'
 
 // Fragment: retrieve single team member fields.
