+++ lib/wordpress/teams/queryTeamsArchive.js
@@ -1,7 +1,7 @@
-import archiveData from '@/lib/wordpress/_partials/archiveData'
-import defaultPageData from '@/lib/wordpress/_partials/defaultPageData'
-import featuredImagePostFields from '@/lib/wordpress/_partials/featuredImagePostFields'
-import globalPostFields from '@/lib/wordpress/_partials/globalPostFields'
+import archiveData from '@/lib/wordpress/_query-partials/archiveData'
+import defaultPageData from '@/lib/wordpress/_query-partials/defaultPageData'
+import featuredImagePostFields from '@/lib/wordpress/_query-partials/featuredImagePostFields'
+import globalPostFields from '@/lib/wordpress/_query-partials/globalPostFields'
 import {gql} from '@apollo/client'
 
 // Fragment: retrieve archive team fields.
